objects: split into separate files
