+++ tests/unit/objects/test_tree.py
@@ -3,7 +3,7 @@ from operator import itemgetter
 import pytest
 
 from dvc.hash_info import HashInfo
-from dvc.objects import HashFile
+from dvc.objects.file import HashFile
 from dvc.objects.tree import Tree, _merge
